proyecto_medico/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.7.0
oracledb==3.1.1
gunicorn
